# docs/swagger.yaml
basePath: /
definitions:
  handler.CreateWebhookRequest:
    properties:
      direction:
        enum:
        - above
        - below
        example: above
        type: string
      from:
        example: USD
        type: string
      threshold:
        example: 5.8
        type: number
      to:
        example: BRL
        type: string
      url:
        example: https://example.com/notify
        type: string
    required:
    - direction
    - from
    - threshold
    - to
    - url
    type: object
  handler.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  repository.RateRecord:
    properties:
      cached:
        type: boolean
      from:
        type: string
      id:
        type: integer
      queried_at:
        type: string
      rate:
        type: number
      to:
        type: string
    type: object
  repository.Webhook:
    properties:
      created_at:
        type: string
      direction:
        type: string
      from:
        type: string
      id:
        type: integer
      last_triggered_at:
        type: string
      threshold:
        type: number
      to:
        type: string
      url:
        type: string
    type: object
  service.RateResult:
    properties:
      cached:
        type: boolean
      from:
        type: string
      rate:
        type: number
      to:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Currency exchange rate API with Redis caching, PostgreSQL history,
    rate limiting, and webhooks.
  title: Exchange Rate API
  version: "1.0"
paths:
  /health:
    get:
      description: Returns the status of the API and its dependencies
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check
      tags:
      - system
  /rates:
    get:
      description: Returns the current exchange rate between two currencies. Served
        from Redis cache when available.
      parameters:
      - description: Source currency code (e.g. USD)
        in: query
        name: from
        required: true
        type: string
      - description: Target currency code (e.g. BRL)
        in: query
        name: to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.RateResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get exchange rate
      tags:
      - rates
  /rates/history:
    get:
      description: Returns historical exchange rate records for a currency pair stored
        in the database
      parameters:
      - description: Source currency code (e.g. USD)
        in: query
        name: from
        required: true
        type: string
      - description: Target currency code (e.g. BRL)
        in: query
        name: to
        required: true
        type: string
      - description: 'Number of past days to include (default: 7)'
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repository.RateRecord'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get rate history
      tags:
      - rates
  /webhooks:
    get:
      description: Returns all registered webhooks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repository.Webhook'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: List webhooks
      tags:
      - webhooks
    post:
      consumes:
      - application/json
      description: Registers a URL to be notified when an exchange rate crosses a
        threshold
      parameters:
      - description: Webhook configuration
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateWebhookRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/repository.Webhook'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Register a webhook
      tags:
      - webhooks
  /webhooks/{id}:
    delete:
      description: Removes a registered webhook by ID
      parameters:
      - description: Webhook ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Delete a webhook
      tags:
      - webhooks
swagger: "2.0"